HelloWorld cleanup and additions.
Removed reference to Release Candidate from VS2015.
Changed starting structure to sub bullets to keep readability and flow
of intro.

Changed Universal Windows App reference to Universal Windows Platform
as the rest of the guide calls it that.

Removed lines of block text (>) so that when rendered it shows as a
single flowing section and not separate blocks.

Removed number list from adding content to main page. It read as a
conversation instead of explicitly a bullet list.

Added link out to learn more on UWP’s.

Changed VS to Visual Studio as rest of guide uses full name.