removed 'Max' from the Minnowboard button
The Minnowboard OS image works with ALL the MinnowBoards (Max, Turbot etc).  Removing 'Max' will be clearer for users to choose Minnowboard for all their OS install.  It also helps in other instructions off-site to not have to mention to use "Max" even though you have a "Turbot" etc.  This helps with Usability and User Experience.