Update SuggestedDevices.md
"After review it seems hardware changes in the RPi2 v1.2 make it incompatible with the FFU created by Microsoft for the original RPi2 and new RPi3. Microsoft is not developing changes to support the Pi2 V1.2 hardware and recommends using the more popular RPi3. 

However, the existing drivers for Raspberry Pi2 and RPi3 are available at https://github.com/ms-iot/bsp for individuals interested in investigating developing their own image for Rpi2 v1.2."